O ponto ganha aba própria, e a primeira batida real não entrava
DOIS BUGS, e o segundo escondeu o primeiro por uma hora.

1. `precisao` estava como `z.number().int()` — o único `precisao` do
   projeto assim. `Location.getCurrentPositionAsync` devolve `accuracy`
   float (13.456), então TODA marcação com GPS era recusada, que é o
   caminho normal. Agora aceita float e o servidor arredonda na hora de
   gravar na coluna Int.

2. O app humanizava QUALQUER `invalid_type` como "obrigatório (não foi
   enviado)". O motorista (e eu) fomos procurar um campo em branco que
   estava preenchido. Agora distingue "não veio" de "veio no formato
   errado", usando a mensagem do próprio Zod.

E A UI. O dono abriu e disse, com razão: "que porcaria de UI é isso". O
erro não era de estilo, era de LUGAR. Pra quem é registrado em carteira,
bater o ponto é o que ele abre o app pra fazer — três, quatro vezes por
dia, todo dia. Isso é uma ABA, não um card espremido entre catorze
blocos de uma home que é de outro assunto.

A aba "Ponto" aparece só pra quem é funcionário (`href: null` tira do tab
bar E fecha a rota — esconder o botão e deixar o caminho aberto não é
esconder nada). Dentro dela: o relógio grande dizendo QUE HORAS vão ser
gravadas antes do toque, o botão, as batidas do dia em ordem e sem
rótulo, o estado do envio, e os caminhos pro espelho e pra correção.

O card saiu das duas homes.

+++ packages/shared-types/src/ponto.ts
@@ -89,7 +89,14 @@ export const MarcacaoPontoInput = z.object({
   agoraNoAparelho: ISO.optional(),
   latitude: z.number().min(-90).max(90).optional(),
   longitude: z.number().min(-180).max(180).optional(),
-  precisao: z.number().int().min(0).max(100000).optional(),
+  /**
+   * ⚠️ NÃO é inteiro. `Location.getCurrentPositionAsync` devolve `accuracy`
+   * como float (13.456), e um `.int()` aqui recusa TODA marcação com GPS —
+   * que é o caminho normal. Foi o primeiro erro real no app. Todos os outros
+   * `precisao` do projeto já aceitam float; este era o único fora do padrão.
+   * Quem arredonda é o servidor, na hora de gravar na coluna Int.
+   */
+  precisao: z.number().min(0).max(100000).optional(),
   appVersao: z.string().max(40).optional(),
   dispositivo: z.string().max(120).optional(),
 });
